Add UTs for RewriteUnaryNotExprVisitor and fix nested situation
